[UI] NoItemView -> IconMessageView로 이름 변경, 제약 및 생성자 등 수정
